requirements.txt
google-play-scraper>=1.2.7
gspread>=6.0.0
google-auth>=2.28.0
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
rapidfuzz>=3.0.0
